oh-my-zsh installation

+++ install.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-DIR=$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )
+DIR=$(cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd)
 
 # install pathogen
 mkdir -p $DIR/vim/autoload $DIR/vim/bundle
@@ -8,11 +8,20 @@ curl -LSso $DIR/vim/autoload/pathogen.vim https://tpo.pe/pathogen.vim
 
 # compile modules
 git submodule update --init --recursive
-./vim/bundle/YouCompleteMe/install.sh --clang-completer --gocode-completer
-cd ./vim/bundle/vimproc.vim && make
+$DIR/vim/bundle/YouCompleteMe/install.sh --clang-completer --gocode-completer
+cd $DIR/vim/bundle/vimproc.vim && make
+
+# create folders
+mkdir -p ~/.config
+
+# link folders
+[ ! -L ~/.vim ] && ln -s $DIR/vim ~/.vim
+[ ! -L ~/.config/base16-shell ] && ln -s $DIR/base16-shell ~/.config/base16-shell
+[ ! -L ~/.oh-my-zsh ] && ln -s $DIR/oh-my-zsh ~/.oh-my-zsh
+[ ! -L ~/.omz-custom ] && ln -s $DIR/omz-custom ~/.omz-custom
 
 # link scripts
-ln -s $DIR/vim ~/.vim
-ln -s $DIR/vimrc ~/.vimrc
-ln -s $DIR/gitconfig ~/.gitconfig
-ln -s $DIR/gitignore ~/.gitignore
+[ ! -L ~/.zshrc ] && ln -s $DIR/zshrc ~/.zshrc
+[ ! -L ~/.vimrc ] && ln -s $DIR/vimrc ~/.vimrc
+[ ! -L ~/.gitconfig ] && ln -s $DIR/gitconfig ~/.gitconfig
+[ ! -L ~/.gitignore ] && ln -s $DIR/gitignore ~/.gitignore
